Update export/import annotations/backups
